feat(slice-005): tab bar + empty Portfolio tab
- Refactor AppModel into thin root/router model with tab switching
- Extract MarketsModel with all markets-specific state and logic
- Add PortfolioModel with empty state placeholder
- Implement tab bar rendering with active/inactive styling
- Add Tab/Shift+Tab/1/2 for tab navigation
- Add global q/Ctrl+C quit handling
- Suppress tab switching when child input is active (InputActive method)
- Propagate WindowSizeMsg to children with height-1 for tab bar
- Add comprehensive tests for all new functionality
